+++ samples/2d_sample.cpp
@@ -34,7 +34,7 @@ private:
         return true;
     }
 
-    void do_step(float dt) {}
+    void do_step(double dt) {}
     void do_cleanup() {}
 
     Sprite::ptr sprite_;
